@@ -5,15 +5,8 @@ import ConduitLayoutHeader from "./conduit-layout-header";
 import ConduitLayoutFooter from "./conduit-layout-footer";
 
 import ConduitPagesHome from "./conduit-pages-home";
-import { fetchArticles } from "./fetchArticles.http";
-import { fetchTags } from "./fetchTags.http";
 
-import {
-  init,
-  onFeedSelected,
-  onTagSelected,
-  onPageSelected,
-} from "./conduit-pages-home/service";
+const HomePage = ConduitPagesHome();
 
 function App() {
   return (
@@ -21,20 +14,7 @@ function App() {
       <ConduitLayoutHeader />
       <BrowserRouter>
         <Routes>
-          <Route
-            path="/"
-            element={
-              <ConduitPagesHome
-                init={init({
-                  fetchArticles: fetchArticles(),
-                  fetchTags: fetchTags(),
-                })}
-                onFeedSelected={onFeedSelected}
-                onTagSelected={onTagSelected}
-                onPageSelected={onPageSelected}
-              />
-            }
-          />
+          <Route path="/" element={<HomePage />} />
           <Route path="/logout" element={<Navigate replace to="/" />} />
         </Routes>
       </BrowserRouter>
