@@ -2,69 +2,98 @@ import React, { useState, useEffect } from "react";
 import ConduitTagsPopular from "../components/conduit-tags-popular";
 import ConduitArticlesList from "../components/conduit-articles-list";
 import ConduitArticlesFeed from "../components/conduit-articles-feed";
+import {
+  init as serviceOnInit,
+  onFeedSelected as serviceOnFeedSelected,
+  onTagSelected as serviceOnTagSelected,
+  onPageSelected as serviceOnPageSelected,
+} from "./service";
 
-const Home = ({ init, onTagSelected, onFeedSelected, onPageSelected }) => {
-  const [state, setState] = useState();
+import { fetchArticles } from "../fetchArticles.http";
+import { fetchTags } from "../fetchTags.http";
 
-  useEffect(() => {
-    init().then(setState);
-  }, []);
+export default function component(dependencies) {
+  !dependencies && (dependencies = {});
+  return function ({
+    init = serviceOnInit({
+      fetchArticles: dependencies.fetchArticles || fetchArticles(),
+      fetchTags: dependencies.fetchTags || fetchTags(),
+    }),
+    onTagSelected = serviceOnFeedSelected({
+      fetchArticles: dependencies.fetchArticles || fetchArticles(),
+      fetchTags: dependencies.fetchTags || fetchTags(),
+    }),
+    onFeedSelected = serviceOnTagSelected({
+      fetchArticles: dependencies.fetchArticles || fetchArticles(),
+      fetchTags: dependencies.fetchTags || fetchTags(),
+    }),
+    onPageSelected = serviceOnPageSelected({
+      fetchArticles: dependencies.fetchArticles || fetchArticles(),
+      fetchTags: dependencies.fetchTags || fetchTags(),
+    }),
+  }) {
+    const [state, setState] = useState();
 
-  const onTagIsSelected = (tag) => {
-    onTagSelected({ tag, state: getState() }).then(setState);
-  };
+    useEffect(() => {
+      init().then((state) => setState(state));
+    }, []);
 
-  const onFeedIsSelected = (feed) => {
-    onFeedSelected({ feed, state: getState() }).then(setState);
-  };
+    const onTagIsSelected = (tag) => {
+      onTagSelected({ tag, state: getState() }).then(setState);
+    };
 
-  const getState = () => {
-    return JSON.parse(JSON.stringify(state));
-  };
+    const onFeedIsSelected = (feed) => {
+      onFeedSelected({ feed, state: getState() }).then(setState);
+    };
 
-  return (
-    <div className="home-page">
-      <div className="banner">
-        <div className="container">
-          <h1 className="logo-font">conduit</h1>
-          <p>
-            A place to share your <i>Angular</i> knowledge.
-          </p>
-        </div>
-      </div>
-      <div className="container page">
-        <div className="row">
-          <div className="col-md-9">
-            {state && state.feeds ? (
-              <ConduitArticlesFeed
-                feeds={state.feeds}
-                selected={state.selectedFeed}
-                onSelected={onFeedIsSelected}
-              ></ConduitArticlesFeed>
-            ) : (
-              <div>Loading... </div>
-            )}
+    const getState = () => {
+      return JSON.parse(JSON.stringify(state));
+    };
 
-            {state && state.articles ? (
-              <ConduitArticlesList articles={state.articles} />
-            ) : (
-              <div>Loading...</div>
-            )}
-          </div>
-          <div className="col-md-3">
-            {state && state.tags ? (
-              <ConduitTagsPopular
-                tags={state.tags}
-                onSelected={onTagIsSelected}
-              />
-            ) : (
-              <div>Loading...</div>
-            )}
+    return state ? (
+      <div className="home-page">
+        <div className="banner">
+          <div className="container">
+            <h1 className="logo-font">conduit</h1>
+            <p>
+              A place to share your <i>Angular</i> knowledge.
+            </p>
           </div>
         </div>
-      </div>
-    </div>
-  );
-};
+        <div className="container page">
+          <div className="row">
+            <div className="col-md-9">
+              {state["articles.feeds.data"] ? (
+                <ConduitArticlesFeed
+                  feeds={state["articles.feeds.data"]}
+                  selected={state["articles.feeds.selected"]}
+                  onSelected={onFeedIsSelected}
+                ></ConduitArticlesFeed>
+              ) : (
+                <div>Loading... </div>
+              )}
 
-export default Home;
+              {state["articles.list.data"] ? (
+                <ConduitArticlesList articles={state["articles.list.data"]} />
+              ) : (
+                <div>Loading...</div>
+              )}
+            </div>
+            <div className="col-md-3">
+              {state["tags.data"] ? (
+                <ConduitTagsPopular
+                  tags={state["tags.data"]}
+                  onSelected={onTagIsSelected}
+                />
+              ) : (
+                <div>Loading...</div>
+              )}
+            </div>
+          </div>
+        </div>
+      </div>
+    ) : (
+      <div>Loading... </div>
+    );
+  };
+}
